feat: rename forms residence registration
