+++ app/components/forms-residence-registration/first-registration/has-child.tsx
@@ -1,19 +1,18 @@
-import { useFirstRegistrationStore } from "./store";
-import { useProgressStore } from "../../steps/store";
 import { RadioInput } from "../../radio-input";
+import { useFirstRegistrationStore } from "./store";
+import { useProgressStore } from "../../steps-residence-registration/store";
 import { SecondaryButton } from "../../buttons/secondary-button";
 import { i18n } from "~/i18n/i18n-utils";
 import { FormButtonNext } from "~/components/buttons/form-button-next";
 
-export function IsRegisteringChild() {
-	const { isRegisteringChild, setIsRegisteringChild } =
-		useFirstRegistrationStore();
+export function HasChild() {
+	const { hasChild, setHasChild } = useFirstRegistrationStore();
+
+	const isValid = hasChild !== null;
 
 	const { goToPreviousStep, goToNextStep } = useProgressStore();
 
 	const options = ["yes", "no"] as const;
-
-	const isValid = isRegisteringChild !== null;
 
 	return (
 		<form
@@ -25,16 +24,17 @@ export function IsRegisteringChild() {
 		>
 			<div className="flex flex-col gap-4">
 				<h2 className="text-xl font-bold lg:text-4xl">
-					{i18n("first-registration.q5")}
+					{i18n("first-registration.q4")}
 				</h2>
+
 				<div className="flex flex-col gap-1">
 					{options.map((option) => {
-						const name = "first-registration.q5.radio";
+						const name = "first-registration.q4.radio";
 						const label = i18n(option);
 						const isChecked =
-							(option === "yes" && isRegisteringChild === true) ||
-							(option === "no" && isRegisteringChild === false);
-						const onChange = () => setIsRegisteringChild(option === "yes");
+							(option === "yes" && hasChild === true) ||
+							(option === "no" && hasChild === false);
+						const onChange = () => setHasChild(option === "yes");
 
 						return (
 							<RadioInput
