+++ app/components/forms-residence-registration/first-registration/is-married.tsx
@@ -1,22 +1,21 @@
-import { useNationalityStore } from "./store";
-import { useProgressStore } from "../../steps/store";
+import { useFirstRegistrationStore } from "./store";
+import { useProgressStore } from "../../steps-residence-registration/store";
 import { RadioInput } from "../../radio-input";
 import { SecondaryButton } from "../../buttons/secondary-button";
 import { i18n } from "~/i18n/i18n-utils";
 import { FormButtonNext } from "~/components/buttons/form-button-next";
 
-export function IsGerman() {
-	const { isGerman, setIsGerman } = useNationalityStore();
-
-	const isValid = isGerman !== null;
-
+export function IsMarried() {
+	const { isMarried, setIsMarried } = useFirstRegistrationStore();
 	const { goToPreviousStep, goToNextStep } = useProgressStore();
 
 	const options = ["yes", "no"] as const;
 
+	const isValid = isMarried !== null;
+
 	return (
 		<form
-			className="flex h-80 w-full flex-col gap-12 lg:h-96"
+			className="flex w-full flex-col gap-12"
 			onSubmit={(e) => {
 				e.preventDefault();
 				goToNextStep();
@@ -24,16 +23,16 @@ export function IsGerman() {
 		>
 			<div className="flex flex-col gap-4">
 				<h2 className="text-xl font-bold lg:text-4xl">
-					{i18n("nationality.q1")}
+					{i18n("first-registration.q2")}
 				</h2>
 				<div className="flex flex-col gap-1">
 					{options.map((option) => {
-						const name = "nationality.q1.radio";
-						const isChecked =
-							(option === "yes" && isGerman === true) ||
-							(option === "no" && isGerman === false);
-						const onChange = () => setIsGerman(option === "yes");
+						const name = "first-registration.q2.radio";
 						const label = i18n(option);
+						const isChecked =
+							(option === "yes" && isMarried === true) ||
+							(option === "no" && isMarried === false);
+						const onChange = () => setIsMarried(option === "yes");
 
 						return (
 							<RadioInput
