+++ app/components/forms-residence-registration/first-registration/store/index.ts
