+++ app/components/forms-residence-registration/nationality/index.tsx
@@ -1,5 +1,5 @@
-export * from "./is-german";
+export * from "./is-german.tsx";
 export * from "./is-german-over-16.tsx";
-export * from "./is-european";
+export * from "./is-european.tsx";
 export * from "./is-non-german-over-16.tsx";
 export * from "./is-refugee.tsx";
