+++ app/components/forms-residence-registration/nationality/is-non-german-over-16.tsx
@@ -1,21 +1,22 @@
-import { useFirstRegistrationStore } from "./store";
-import { useProgressStore } from "../../steps/store";
 import { RadioInput } from "../../radio-input";
+import { useNationalityStore } from "./store";
+import { useProgressStore } from "../../steps-residence-registration/store";
 import { SecondaryButton } from "../../buttons/secondary-button";
 import { i18n } from "~/i18n/i18n-utils";
 import { FormButtonNext } from "~/components/buttons/form-button-next";
 
-export function IsMarried() {
-	const { isMarried, setIsMarried } = useFirstRegistrationStore();
+export function IsNonGermanOver16() {
+	const { isNonGermanOver16, setIsNonGermanOver16 } = useNationalityStore();
+
+	const isValid = isNonGermanOver16 !== null;
+
 	const { goToPreviousStep, goToNextStep } = useProgressStore();
 
 	const options = ["yes", "no"] as const;
 
-	const isValid = isMarried !== null;
-
 	return (
 		<form
-			className="flex w-full flex-col gap-12"
+			className="flex h-80 w-full flex-col gap-12 lg:h-96"
 			onSubmit={(e) => {
 				e.preventDefault();
 				goToNextStep();
@@ -23,16 +24,16 @@ export function IsMarried() {
 		>
 			<div className="flex flex-col gap-4">
 				<h2 className="text-xl font-bold lg:text-4xl">
-					{i18n("first-registration.q2")}
+					{i18n("nationality.q4")}
 				</h2>
 				<div className="flex flex-col gap-1">
 					{options.map((option) => {
-						const name = "first-registration.q2.radio";
-						const label = i18n(option);
+						const name = "nationality.q4.radio";
 						const isChecked =
-							(option === "yes" && isMarried === true) ||
-							(option === "no" && isMarried === false);
-						const onChange = () => setIsMarried(option === "yes");
+							(option === "yes" && isNonGermanOver16 === true) ||
+							(option === "no" && isNonGermanOver16 === false);
+						const onChange = () => setIsNonGermanOver16(option === "yes");
+						const label = i18n(option);
 
 						return (
 							<RadioInput
