+++ app/components/forms-residence-registration/overview/index.tsx
@@ -1,11 +1,11 @@
-import { useOverviewStore } from "./store";
-import { useProgressStore } from "../../steps/store";
+import { useOverviewStore } from "./store/index.ts";
+import { useProgressStore } from "../../steps-residence-registration/store/index.ts";
 import { DocumentCheckbox } from "./document-checkbox.tsx";
-import { SecondaryButton } from "../../buttons/secondary-button";
+import { SecondaryButton } from "../../buttons/secondary-button/index.tsx";
 import { i18n } from "~/i18n/i18n-utils.ts";
 import { trackInteraction } from "../../feedback/matomo.ts";
 import { useTrackGoalChecklistConversion } from "./hooks/use-track-goal-checklist-conversion.tsx";
-import { PrimaryButton } from "../../buttons/primary-button";
+import { PrimaryButton } from "../../buttons/primary-button/index.tsx";
 
 export function Overview() {
 	const { docs: requiredDocs } = useOverviewStore();
