+++ app/components/forms-residence-registration/overview/store/index.ts
@@ -99,7 +99,7 @@ export const useOverviewStore = create<OverviewStore>()(
 		}),
 
 		{
-			name: "overview",
+			name: "overview-residence-registration",
 		},
 	),
 );
