+++ app/components/steps-residence-registration/store/steps.ts
